requirements.txt
python-dotenv==0.21.0
oracledb==2.4.1